requirements.txt
discord.py[voice]==2.6.3
python-dotenv==1.0.0
pydub==0.25.1
psutil==5.9.0
